module_09/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.4
django-environ==0.8.1
django-phonenumber-field==6.1.0
phonenumbers==8.12.48
sqlparse==0.4.2
